Préparation et nettoyages en vue d'une release "stable". - Correction totale du bug (https://github.com/abotzung/foguefi/issues/31) - Ajout de plugins "apk" pour FOS (https://github.com/abotzung/foguefi/issues/18)   -> Créer un dossier dans /images/@apk et y copier les fichiers apk à l'intérieur. - Clôture de (https://github.com/abotzung/foguefi/issues/16), corrigée dans les plugins apk - TODO : Le thème de GRUB est à réecrire ; Redo Rescue exclut la réutilisation de leurs graphismes (NON GPL : https://github.com/redorescue/redorescue/blob/master/overlay/image/boot/grub/theme/LICENSE)   C'était une copie à l'arrache et j'avais aucune idée de ce que je faisait. Je dois nettoyer tout cela.

+++ install.sh
@@ -1,28 +1,26 @@
 #!/bin/bash
 #============================================================================
-#              F O G    P R O J E C T    v 1 . 5 . x
-#                    Unofficial Secure Boot Patch
+#         F O G U E F I - Free Opensource Ghost, batteries included
+# An unofficial portage of GRUB and FOS for an easy useage of FOG Server on 
+#                      Secure Boot enabled computers.
+# 
 #             FOGUefi (https://github.com/abotzung/foguefi)
 #
-# Auteur       : Alexandre BOTZUNG [alexandre.botzung@grandest.fr]
-# Auteur       : The FOG Project team (https://github.com/FOGProject/fogproject)
-# Version      : 20240321
+# Author       : Alexandre BOTZUNG [alexandre.botzung@grandest.fr]
+# Author       : The FOG Project team (https://github.com/FOGProject/fogproject)
+# Version      : 20240808
 # Licence      : http://opensource.org/licenses/gpl-3.0
 #============================================================================ 
 # install.sh
-#   Ce script déploie le patch foguefi sur le système.
+#   This script deploy FOGUefi on this system.
 #
-#   Celui-ci est composée en 3 parties : 
-#    - GRUB et SHIM (déployés dans /tftpboot) 
-#    - Divers fichiers php (déployés dans /var/www/?/fog)
-#    - Le chroot bureau graphique (déployé dans /images) (absent dans cette version)
+#   This script :
+#    - Download *or* compiles a FOS Client 
+#    - Deploy Linux kernel, GRUB and SHIM (latest signed) into /tftpboot
+#    - Deploy GrubBootMenu php files (copied into /var/www/?/fog)
+#    - If FOG works on HTTPS, reconfigure Apache2.
 #============================================================================
 
-# TODO FIXME : The english language used in this installer (BuilfFogUefi included) is
-#    at best approximate. A complete cleaning is in order!
-#
-# TODO (FOG Version : https://raw.githubusercontent.com/FOGProject/fogproject/master/packages/web/lib/fog/system.class.php )
-#
 
 # Oh ! Dirty !  ;
 source /opt/fog/.fogsettings
@@ -51,17 +49,7 @@ echo ' > This patch extends the FOG PXE boot possibility to Secure Boot enabled 
 echo '   It consists of 3 parts : '
 echo '   - Files required for PXE (shim/GRUB), and FOG Stub patched'
 echo '   - PHP Files for handling newer menus for GRUB'
-echo '   - An optionnal tool : X Server for FOG Stub, allowing to surfing/remote desktop while FOG is working'
 echo ''
-echo ' This patch contains files from :'
-echo ' - The FOG Project <https://fogproject.org/> (init.xz & scripts & logos)'
-echo ' - Clonezilla (C) 2003-2023, NCHC, Taiwan <https://clonezilla.org/> (scripts)'
-echo ' - Ubuntu (C) 2023 Canonical Ltd. <https://ubuntu.com/> (GNU/Linux signed kernel, shim-signed, grub-efi-arm64-signed)'
-echo ' - The Alpine Linux Development team <https://www.alpinelinux.org/> (base env)'
-echo ' - Redo Rescue (C) 2010.2020 Zebradots Software <http://redorescue.com/> (GRUB Theme, heavily modified)'
-echo ' - Mcder3 <github.com/KaOSx/midna> (icons)'
-echo ' - Gnome icon pack <https://download.gnome.org/sources/gnome-icon-theme/> (icons) (c) 2002-2008 :'
-echo '  '
 echo ''
 echo 'This patch are free software; the exact distribution terms for each program are described in the individual files.'
 echo 'This patch comes with ABSOLUTELY NO WARRANTY, to the extent permitted by applicable law.'
@@ -102,27 +90,27 @@ if [[ "$question" == "y" || "$question" == "Y" ]]; then
 	cd "$basedir" || exit
 
 
-	echo "=> Copying GRUB files..."
+	echo "=> Copy GRUB files..."
 	cp -rf "$basedir"/src/tftpboot/* /tftpboot/
 	chown -R fogproject:root /tftpboot
 	chmod -R 0755 /tftpboot
 	# DEBUGME DEBUGME - Switched to no delay for dev purposes
 	#sleep 1
 
-	echo "=> Copying FOG PHP files..."
+	echo "=> Copy FOG PHP files..."
 	cp -rf "$basedir"/src/fog/* "${docroot}${webroot}"
 	chown -R www-data:www-data "${docroot}${webroot}"
 	chmod -R 0755 "${docroot}${webroot}/lib"
 	chmod -R 0755 "${docroot}${webroot}/service"
-	rm "${docroot}${webroot}/service/grub/tftp"
-	rm "${docroot}${webroot}/service/grub/grub_https.php"
+	[[ -L "${docroot}${webroot}/service/grub/tftp" && -d "${docroot}${webroot}/service/grub/tftp" ]] && rm "${docroot}${webroot}/service/grub/tftp"
+	[[ -r "${docroot}${webroot}/service/grub/grub_https.php" ]] && rm "${docroot}${webroot}/service/grub/grub_https.php"
 	ln -s /tftpboot "${docroot}${webroot}/service/grub/tftp"
 	ln -s "${docroot}${webroot}/service/grub/grub.php" "${docroot}${webroot}/service/grub/grub_https.php"
 	chmod +x "${docroot}${webroot}/service/grub/grub_https.php"
 	# DEBUGME DEBUGME - Switched to no delay for dev purposes
 	#sleep 1
 
-    echo "=> Configurying Apache server..."
+    echo "=> Configure Apache server..."
     FOGApacheFile=$(grep -rnw '/management/other/ca.cert.der$ - ' /etc/apache2 | head -n1 | cut -f1 -d:)
     FOGApachefileAlreadyPatched=$(grep -rnw 'Add made by foguefi patch' /etc/apache2 | head -n1 | cut -f1 -d:)
     if [ -f "$FOGApacheFile" ]; then
@@ -156,7 +144,7 @@ EOF
     
 	echo "==== Installation done ! ===="
 	echo ''
-	echo " REMEMBER to change your PXE Boot file $bootfilename to shimx64.efi in your DHCP Server (option 67)"
+	echo -e " REMEMBER to change your PXE Boot file  $bootfilename to \033[30;42m shimx64.efi \033[0m in your DHCP Server (option 67)"
 	echo ''
 	echo ' - Have a nice day !'
